refactor(src): api 디렉터리의 index에서 export 및 변경사항 반영
