+++ frontend/src/api/index.ts
@@ -1,4 +1,11 @@
 export { CATEGORY } from './category';
 export { customFetch } from './customFetch';
 export { QUERY_KEY } from './queryKeys';
-export { TEMPLATE_API_URL, getTemplateList, getTemplate, postTemplate } from './templates';
+export {
+  TEMPLATE_API_URL,
+  getTemplateList,
+  getTemplate,
+  postTemplate,
+  editTemplate,
+  deleteTemplate,
+} from './templates';
